requirements.txt
fastapi~=0.115.12
pydantic~=2.11.7
openai~=1.86.0
python-dotenv~=1.1.0
SQLAlchemy~=2.0.41
alembic~=1.16.2
asyncpg~=0.30.0
together~=1.5.16
feedparser~=6.0.11
requests~=2.32.4
beautifulsoup4~=4.13.4
passlib~=1.7.4